discord messages + bugfixes

+++ Discord2Polygon/SendToPolygonDTO.cs
@@ -7,11 +7,11 @@ namespace Discord2Polygon;
 public class SendToPolygonDTO : FunctionMessage
 {
     [Parameter("uint256", "syncSumDiscordBalances", 1)]
-    public int syncSumDiscordBalances { get; set; }
+    public long syncSumDiscordBalances { get; set; }
         
     [Parameter("address", "addressOfDiscordUser", 2)]
     public string addressOfDiscordUser { get; set; }
         
     [Parameter("uint256", "payoutAmount", 3)]
-    public int payoutAmount { get; set; }
+    public long payoutAmount { get; set; }
 }
